- Resolved open issue #3 (Unable to build under Debian 11/Bullseye) - Removed an override of libgmime version forcing V2 to be recognized - Tested that both the ax25 and telnet transports retrieve mail successfully - Added a missing socket close() which was leaking open sockets - Doubled the child process timeout to fix spurious timeouts
The telnet and AX25 retrieval methods are tested and seem to be working
properly now. To my understanding, it shouldn't be possible to leak
sockets into the operating system because they should close when
the process exits, but that's nonetheless what I was seeing. It was
very detrimental to be able to run the program only once before
you were stuck with an address-busy error and had to restart the
radio stack. That's fixed. It might be a good idea to add signal handlers
to keep the connection from dangling on SIGINT, for example.