Resolved Issue #3 (Cannot build with libgmime 3 or Debian 11/Bullseye)
- Removed the override which forces gmime version detection to v2
- Fixed the transfer of the message date between Winlink and MIME header for gmime 3
- That seemed to be the last thing missing for gmime 3 to work, but I have not tested the serial transport
- Doubled the fork process timeout to avoid spurious timeouts
- Added a missing close() call in the forked process which appeared to be leaking sockets into the operating system

The last item is especially weird, since I don't think it's supposed to be possible; the OS is supposed to close
all file handles (including sockets) upon process exit, so maybe it represents a kernel or driver bug. Anyway, I
find it a valuable workaround, as I would otherwise have to restart the radio stack after every run.
This leak may happen anyway if the program closes abnormally, but at least it's fixed for the more common cases.
If it persists, we might want to put the socket close in a signal or termination handler.